Restructured lifecycle regarding messages
