+++ src/main/java/zork/enviromnment/messages/DeviceLifecycle.java
@@ -1,0 +1,17 @@
+package zork.enviromnment.messages;
+
+public class DeviceLifecycle {
+
+  public static final class RequestTrackDevice {
+    public final String groupId;
+    public final String deviceId;
+
+    public RequestTrackDevice(String groupId, String deviceId) {
+      this.groupId = groupId;
+      this.deviceId = deviceId;
+    }
+  }
+
+  public static final class DeviceRegistered {
+  }
+}
